% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratpack.R
\name{get_user_packages}
\alias{get_user_packages}
\title{returns users packages from outside the packrat library}
\usage{
get_user_packages()
}
\value{
users installed packages
}
\description{
returns users packages from outside the packrat library
}
\examples{
get_user_packages()
}
